Domain layer organization.
