--- app/src/main/java/com/example/rememberme/Board.kt
@@ -1,7 +1,0 @@
-package com.example.rememberme
-
-data class Board (
-    val image: Int,
-    val title: String,
-    val description: String
-)
