+++ app/src/main/java/com/example/rememberme/common/extensions.kt
@@ -1,8 +1,27 @@
 package com.example.rememberme.common
 
 import android.app.Activity
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import androidx.recyclerview.widget.DiffUtil
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
-fun Activity.setBottomNavigationVisibility(bottomNavigation: Int, visibility: Int){
+fun Activity.setBottomNavigationVisibility(bottomNavigation: Int, visibility: Int) {
     findViewById<BottomNavigationView>(bottomNavigation).visibility = visibility
 }
+
+inline fun <T : Any> basicDiffUtil(
+    crossinline areItemsTheSame: (T, T) -> Boolean = { old, new -> old == new },
+    crossinline areContentsTheSame: (T, T) -> Boolean = { old, new -> old == new },
+) = object : DiffUtil.ItemCallback<T>() {
+    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean =
+        areItemsTheSame(oldItem, newItem)
+
+    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean =
+        areContentsTheSame(oldItem, newItem)
+}
+
+fun ViewGroup.inflate(@LayoutRes layoutRes: Int, attachToRoot: Boolean = true): View =
+    LayoutInflater.from(context).inflate(layoutRes, this, attachToRoot)
