--- app/src/main/java/com/example/rememberme/extensions.kt
@@ -1,21 +1,0 @@
-package com.example.rememberme
-
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.recyclerview.widget.DiffUtil
-
-inline fun <T : Any> basicDiffUtil(
-    crossinline areItemsTheSame: (T, T) -> Boolean = { old, new -> old == new },
-    crossinline areContentsTheSame: (T, T) -> Boolean = { old, new -> old == new },
-) = object : DiffUtil.ItemCallback<T>() {
-    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean =
-        areItemsTheSame(oldItem, newItem)
-
-    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean =
-        areContentsTheSame(oldItem, newItem)
-}
-
-fun ViewGroup.inflate(@LayoutRes layoutRes: Int, attachToRoot: Boolean = true): View =
-    LayoutInflater.from(context).inflate(layoutRes, this, attachToRoot)
