+++ domain/src/main/java/com/example/rememberme/domain/Board.kt
@@ -1,0 +1,7 @@
+package com.example.rememberme.domain
+
+data class Board(
+    val image: Int,
+    val title: String,
+    val description: String,
+)
