--- domain/src/main/java/com/example/rememberme/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.example.rememberme.domain
-
-class MyClass {
-}
